requirements.txt
seleniumwire
webdriver-manager
numpy